splited up week1 and week2 files to the seperated files
