+++ hw-week4/react-route/src/components/Week1/DynamicComponent.css
@@ -1,0 +1,3 @@
+.class-div {
+  margin: 1rem;
+}
