+++ hw-week4/react-route/src/components/Week1/Header.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+// This component is header aimed for top of each list
+const Header = ({ name }) => <h1 className="header">{name}</h1>;
+
+export default Header;
