+++ hw-week4/react-route/src/components/Week1/StaticComponent.css
@@ -1,0 +1,3 @@
+.class-div {
+  margin: 1rem;
+}
