+++ hw-week4/react-route/src/components/Week2/Button.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Button = ({ text, clickHandler }) => {
+  return <button onClick={clickHandler}>{text}</button>;
+};
+
+export default Button;
