+++ hw-week4/react-route/src/components/Week2/ItemField.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const ItemField = ({ formHandler }) => {
+  return (
+    <form onSubmit={formHandler}>
+      <input type="text" name="itemField" />
+      <button type="submit">Add Todo</button>
+    </form>
+  );
+};
+
+export default ItemField;
